+++ app/__main__.py
@@ -5,7 +5,10 @@ from openpyxl import load_workbook
 from openpyxl.worksheet.worksheet import Worksheet
 
 from .ecs2terraform import Ecs2Terraform
+from .eip2terraform import Eip2Terraform
 from .enterpriseproj2terraform import EnterpriseProj2Terraform
+from .nat2terraform import Nat2Terraform
+from .natrule2terraform import NatRule2Terraform
 from .secgroup2terraform import Secgroup2Terraform
 from .subnet2terraform import Subnet2Terraform
 from .utils import load_metadata, load_sheet_data
@@ -79,6 +82,41 @@ def process_secgroup(worksheet: Worksheet):
         secgroup_handler.output_terraform_code(output_file)
 
 
+def process_eip(worksheet: Worksheet):
+    data = load_sheet_data(worksheet)
+
+    eip_handler = Eip2Terraform()
+
+    process_sheet_data(worksheet.title, data, eip_handler.add_eip)
+
+    with open('tf/eips.tf', 'w') as output_file:
+        eip_handler.output_terraform_code(output_file)
+
+
+def process_nat(worksheet: Worksheet):
+    data = load_sheet_data(worksheet)
+
+    nat_handler = Nat2Terraform()
+
+    process_sheet_data(worksheet.title, data, nat_handler.append)
+
+    with open('tf/nat.tf', 'w') as output_file:
+        nat_handler.output_terraform_code(output_file)
+
+    return nat_handler.to_dict()
+
+
+def process_nat_rules(worksheet: Worksheet, nat_data: dict):
+    data = load_sheet_data(worksheet)
+
+    rule_handler = NatRule2Terraform(nat_data)
+
+    process_sheet_data(worksheet.title, data, rule_handler.append)
+
+    with open('tf/nat.tf', 'a') as output_file:
+        rule_handler.output_terraform_code(output_file)
+
+
 def process_enterpriseproj(worksheet: Worksheet):
     data = load_sheet_data(worksheet)
 
@@ -109,6 +147,12 @@ def main():
     process_subnet(workbook[metadata['subnet_sheet']])
 
     process_secgroup(workbook[metadata['secgroup_sheet']])
+
+    process_eip(workbook[metadata['eip_sheet']])
+
+    nat_data = process_nat(workbook[metadata['nat_sheet']])
+
+    process_nat_rules(workbook[metadata['nat_rules_sheet']], nat_data)
 
     process_enterpriseproj(workbook[metadata['enterprise_proj_sheet']])