+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+- Add EIP support
+- Add NAT Gateway and NAT rules support
+
 ## [0.3.2] - 2023-09-21
 
 - Add [LLD documentation](docs/LLD.md);